Tests/LibWeb/Text/expected/clip-path-hit-testing.txt

Inside circle: PASS
Outside circle (in bbox): PASS
Inside circle edge: PASS
